Add JungleMap and update game state handling
it now starts in the menu then to the lobby. two tile the teleports the player to two level. waiting to add more when the other levels are ready
the left tile goes to the normal level while the other goes to a test map. the test map as a tile to go back to the lobby